refactor(#86): split HTTP adapter into focused modules
Split internal/adapter/http/graphs.go (894 lines) into seven
focused files within the same package, and extract
d2SourceForPackage from packages_handler.go.

New files:
- graph_payload.go: graphPayload/graphMeta types
- graph_layer.go: buildLayerGraph + layer ordering
- graph_package.go: buildPackageOverviewGraph + label helpers
- graph_diff.go: buildDiffGraph + parent helpers
- export_d2_diff.go: renderDiffD2 + diffOpColor
- export_d2_packages.go: d2SourceForPackage
- export_io.go: writeText/writeSVG/safeFilename

graphs.go retains only HTTP route registration and handler
glue. No behavior change - identical APIs, identical templates,
identical tests pass (go vet ./... clean, go test ./... green).

Closes #86
